SOFTWARE/STROBE_UI/requirements.txt
numpy
pyqt5
pyqtgraph
pyserial
